requirements.txt
# =============================================================================
# AI-POWERED QUIZ GENERATION REQUIREMENTS
# Uncomment the section below for AI-powered quiz generation functionality
# =============================================================================

# Core AI and ML packages
# openai==1.101.0
# google-generativeai==0.8.5
# anthropic==0.64.0
# transformers==4.55.4
# torch==2.5.1+cu121
# torchaudio==2.5.1+cu121
# torchvision==0.17.1+cu121
# torchmetrics==1.3.1
# pytorch-lightning==2.2.5
# lightning==2.2.5
# lightning-cloud==0.5.69
# lightning-utilities==0.11.2
# lightning_sdk==0.1.10
# litdata==0.2.7

# Text-to-Speech and Audio Processing
# TTS==0.22.0
# soundfile==0.13.1
# librosa==0.10.0
# soxr==0.5.0.post1
# audioread==3.0.1
# gruut==2.2.3
# gruut-ipa==0.13.0
# gruut_lang_de==2.0.1
# gruut_lang_en==2.0.1
# gruut_lang_es==2.0.1
# gruut_lang_fr==2.0.2
# phonemizer-fork==3.3.2
# espeakng-loader==0.2.4
# kokoro==0.9.4
# greenlet==3.2.4

# Language Processing and NLP
# spacy==3.8.7
# spacy-curated-transformers==0.3.1
# spacy-legacy==3.0.12
# spacy-loggers==1.0.5
# en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
# nltk==3.9.1
# jieba==0.42.1
# pypinyin==0.55.0
# pysbd==0.3.4
# bangla==0.0.5
# bnnumerizer==0.0.2
# bnunicodenormalizer==0.1.7
# g2pkk==0.1.2
# hangul-romanize==0.1.0
# jamo==0.4.1
# misaki==0.9.4
# SudachiPy==0.6.10
# SudachiDict-core==20250515

# Cloud Storage and File Handling
# aioboto3==13.1.1
# aiobotocore==2.13.1
# boto3==1.34.117
# botocore==1.34.117
# s3transfer==0.10.1
# awscrt==0.20.9
# aiofiles==24.1.0
# cloudpathlib==0.21.0
# fsspec==2024.5.0
# smart_open==7.3.0.post1

# Scientific Computing and Data Processing
# numpy==1.21.6
# scipy==1.11.4
# pandas==1.5.3
# scikit-learn==1.1.0
# numba==0.55.0
# llvmlite==0.38.1
# joblib==1.4.2
# threadpoolctl==3.5.0
# umap-learn==0.5.7
# pynndescent==0.5.13

# CUDA and GPU Support (uncomment if using GPU)
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.0.2.54
# nvidia-cufile-cu12==1.13.1.3
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-cusparselt-cu12==0.7.1
# nvidia-nccl-cu12==2.21.5
# nvidia-nvjitlink-cu12==12.8.93
# nvidia-nvtx-cu12==12.1.105
# triton==3.1.0

# Additional AI/ML Dependencies
# safetensors==0.6.2
# tokenizers==0.21.4
# huggingface-hub==0.34.4
# einops==0.8.1
# encodec==0.1.1
# hf-xet==1.1.8
# curated-tokenizers==0.0.9
# curated-transformers==0.1.1
# trainer==0.0.36
# fire==0.6.0
# coqpit==0.0.17
# loguru==0.7.3
# tqdm==4.66.4
# rich==13.7.1
# typer==0.12.3
# click==8.1.7
# shellingham==1.5.4
# simple-term-menu==1.6.4

# =============================================================================
# MANUAL EXAM UPLOAD & UI REQUIREMENTS
# Uncomment the section below for manual exam upload and Streamlit UI functionality
# =============================================================================

# Web Framework and UI
streamlit==1.40.1
fastapi==0.111.0
fastapi-cli==0.0.4
uvicorn==0.30.1
starlette==0.37.2
python-multipart==0.0.9
greenlet==3.2.4
# Database
sqlalchemy==2.0.43
asyncpg==0.30.0

# Environment and Configuration
python-dotenv==1.0.1

# Data Validation and Serialization
pydantic==2.7.2
pydantic_core==2.18.3

# HTTP and Networking
httpx==0.27.0
httpcore==1.0.5
aiohttp==3.9.5
aioitertools==0.12.0
aiofiles==24.1.0
requests==2.32.3
urllib3==2.3.0

# JSON and Data Processing
orjson==3.10.3
ujson==5.10.0
jsonlines==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
fastjsonschema==2.19.1
jsonpointer==2.4
json5==0.9.25

# Date and Time
python-dateutil==2.9.0
pytz==2024.1
tzdata==2024.1
tzlocal==5.3.1
arrow==1.3.0
isodate==0.7.2
isoduration==20.11.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1

# Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.4
tenacity==9.1.2
backoff==2.2.1
cachetools==5.3.3
filelock==3.14.0
platformdirs==4.2.2
packaging==24.0
typing_extensions==4.14.1
annotated-types==0.7.0
typeguard==4.4.4

# Security and Authentication
PyJWT==2.8.0
cryptography
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0

# Text Processing
beautifulsoup4==4.12.3
lxml
html5lib
bleach==6.1.0
markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
mistune==3.0.2
pygments==2.17.2

# File and Path Handling
pathlib
pathspec
watchdog==6.0.0
watchfiles==0.22.0

# System and Process Management
psutil==6.1.1
distro==1.9.0

# Development and Debugging
debugpy==1.8.1
ipython
jedi==0.19.1
parso
prompt_toolkit==3.0.45
stack_data
traitlets==5.13.0
wcwidth==0.2.11

# Core Python Dependencies
certifi
charset-normalizer==3.4.3
idna==3.10
six==1.16.0
wrapt==1.17.3